# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api backend_challeng_doc88_back",
        "version": "0.1"
    },
    "paths": {
        "/v1/client": {
            "get": {
                "summary": "Paginação para lista de clientes",
                "operationId": "App\\Http\\Controllers\\ClientController::getPagination",
                "parameters": [
                    {
                        "name": "page",
                        "in": "header",
                        "description": "Página a ser exibida",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "header",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "15"
                        },
                        "example": 1
                    },
                    {
                        "name": "order_by",
                        "in": "header",
                        "description": "Campo para ordenação da pesquisa de clientes",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        },
                        "example": "cep"
                    },
                    {
                        "name": "order",
                        "in": "header",
                        "description": "Ordenação da pesquisa asc ou desc",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "asc"
                    },
                    {
                        "name": "ID",
                        "in": "header",
                        "description": "Valor para o campo 'ID'.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "name",
                        "in": "header",
                        "description": "Valor para o campo 'name', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Fulado"
                    },
                    {
                        "name": "email",
                        "in": "header",
                        "description": "Valor para o campo 'email', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@teste.com"
                    },
                    {
                        "name": "address",
                        "in": "header",
                        "description": "Valor para o campo 'address', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Rua dos Tolos"
                    },
                    {
                        "name": "complement",
                        "in": "header",
                        "description": "Valor para o campo 'complement', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ap 25"
                    },
                    {
                        "name": "complement",
                        "in": "header",
                        "description": "Valor para o campo 'neighborhood' (bairro), pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Berrini"
                    },
                    {
                        "name": "phone",
                        "in": "header",
                        "description": "Valor para o campo 'phone', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2345678"
                    },
                    {
                        "name": "cep",
                        "in": "header",
                        "description": "Valor para o campo 'cep', pesquisa do tipo 'like'",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "00000001"
                    },
                    {
                        "name": "birth_date_start",
                        "in": "header",
                        "description": "Valor para o campo 'birth_date', pesquisa a partir desta data",
                        "required": false,
                        "schema": {
                            "type": "date"
                        },
                        "example": "1993-09-22"
                    },
                    {
                        "name": "birth_date_end",
                        "in": "header",
                        "description": "Valor para o campo 'birth_date', pesquisa até esta data",
                        "required": false,
                        "schema": {
                            "type": "date"
                        },
                        "example": "1993-09-22"
                    },
                    {
                        "name": "show_deleted",
                        "in": "header",
                        "description": "Exibir (1) ou não (0) os dados deletados",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                "0",
                                "1"
                            ]
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uma lista de clientes com paginação"
                    },
                    "406": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "summary": "Insere um novo cliente",
                "operationId": "App\\Http\\Controllers\\ClientController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "header",
                        "description": "Valor para o campo 'name'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Fulado"
                    },
                    {
                        "name": "email",
                        "in": "header",
                        "description": "Valor para o campo 'email'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "teste@teste.com"
                    },
                    {
                        "name": "address",
                        "in": "header",
                        "description": "Valor para o campo 'address'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Rua dos Tolos"
                    },
                    {
                        "name": "complement",
                        "in": "header",
                        "description": "Valor para o campo 'complement'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Ap 25"
                    },
                    {
                        "name": "neighborhood",
                        "in": "header",
                        "description": "Valor para o campo 'neighborhood' (bairro)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Berrini"
                    },
                    {
                        "name": "phone",
                        "in": "header",
                        "description": "Valor para o campo 'phone'",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2345678"
                    },
                    {
                        "name": "cep",
                        "in": "header",
                        "description": "Valor para o campo 'cep'",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "00000001"
                    },
                    {
                        "name": "birth_date",
                        "in": "header",
                        "description": "Valor para o campo 'birth_date'",
                        "required": true,
                        "schema": {
                            "type": "date"
                        },
                        "example": "1993-09-22"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Novo cliente registrado com sucesso"
                    },
                    "406": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
...
        "/v1/order/{id}": {
            "get": {
                "summary": "Exibe um pedido ativo pelo {id}",
                "operationId": "App\\Http\\Controllers\\OrderController::getById",
                "parameters": [
                    {
                        "name": "show_deleted",
                        "in": "header",
                        "description": "{id} do pedido a ser localizado.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido solicitado"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "summary": "Alterações no pedido não são autorizadas, você deve cancelá-lo e efetuar um novo pedido.",
                "operationId": "App\\Http\\Controllers\\OrderController::update",
                "responses": {
                    "404": {
                        "description": "Alterações no pedido não são autorizadas, você deve cancelá-lo e efetuar um novo pedido."
                    }
                }
            },
            "delete": {
                "summary": "Deleta um pedido pelo {id}",
                "operationId": "App\\Http\\Controllers\\OrderController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "header",
                        "description": "{id} do pedido a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem informando que o pedido foi deletado."
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados."
                    },
                    "default": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
        "/v1/product/{id}": {
            "get": {
                "summary": "Exibe um produto ativo pelo {id}",
                "operationId": "App\\Http\\Controllers\\ProductController::getById",
                "parameters": [
                    {
                        "name": "show_deleted",
                        "in": "header",
                        "description": "{id} do produto a ser localizado.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do produto solicitado"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "summary": "Deleta um produto pelo {id}",
                "operationId": "App\\Http\\Controllers\\ProductController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "header",
                        "description": "{id} do produto a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem informando que o registro foi deletado."
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados."
                    },
                    "default": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
...
        "/v1/type/all": {
            "get": {
                "summary": "Exibe todos os registros, exceto deletados",
                "operationId": "App\\Http\\Controllers\\TypeController::getAll",
                "responses": {
                    "200": {
                        "description": "Lista de tipo de produtos"
                    },
                    "500": {
                        "description": "Erro no formato ou tipo dos dados enviados"
                    },
                    "default": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
...
    }
}